[issues UI/UX]

+++ src/app/projects/kmmp/actividades/list/activity/activity.component.ts
@@ -1,13 +1,15 @@
 import { Component, Input, OnInit } from "@angular/core";
 
 import { Router } from "@angular/router";
-
+import { MatDialog } from "@angular/material/dialog";
 //SERVICES
 import { ActivitiesService } from "../../activities.service";
 
 //MODELS
 import { ActivityFake } from "app/projects/kmmp/fake-db/activities/activity-fake-db";
 import { EditarFormatoService } from "app/projects/kmmp/formatos/editar-formato/editar-formato.service";
+import { FuseConfirmationService } from "@fuse/services/confirmation";
+import { UiDialogsComponent } from "app/shared/ui/ui-dialogs/ui-dialogs.component";
 
 @Component({
   selector: "app-activity",
@@ -21,7 +23,9 @@ export class ActivityComponent implements OnInit {
   constructor(
     private activitiesService: ActivitiesService,
     private router: Router,
-    private _editarFormatoService: EditarFormatoService
+    private _editarFormatoService: EditarFormatoService,
+    private _fuseConfirmationService: FuseConfirmationService,
+    private matDialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -34,12 +38,44 @@ export class ActivityComponent implements OnInit {
     });
   }
   async removeFormat(index: number) {
-    const formatSelected = await this.activityInfo.formatos[index];
-    this.activitiesService
-      .deleteFormatById(formatSelected.idActividadFormato)
-      .subscribe((resp: any) => {
-        this.activityInfo.formatos.splice(index, 1);
+    if (this.activityInfo.formatos.length > 1) {
+      const dialogRef = this._fuseConfirmationService.open({
+        title: "Eliminar formato",
+        message: "¿Estás seguro que desea eliminar el formato?",
+
+        actions: {
+          confirm: {
+            label: "Sí, eliminar",
+            color: "primary",
+          },
+          cancel: {
+            label: "No",
+          },
+        },
+        dismissible: true,
       });
+
+      dialogRef.beforeClosed().subscribe(async (result) => {
+        const formatSelected = await this.activityInfo.formatos[index];
+
+        if (result === "confirmed") {
+          this.activitiesService
+            .deleteFormatById(formatSelected.idActividadFormato)
+            .subscribe(() => {
+              this.activityInfo.formatos.splice(index, 1);
+            });
+        }
+      });
+    } else {
+      this.matDialog.open(UiDialogsComponent, {
+        width: "500px",
+        data: {
+          title: "Advertencia",
+          message:
+            "Debe tener al menos un formato por informe, no es posible eliminarlo!",
+        },
+      });
+    }
   }
 
   enableDeleteFormats(): boolean {
@@ -57,6 +93,14 @@ export class ActivityComponent implements OnInit {
           this.router.navigate([
             `/admin/actividades/validation/${this.activityData.id}/${e.idActividadFormato}/${secciones.body.secciones[0].id}`,
           ]);
+        } else {
+          this.matDialog.open(UiDialogsComponent, {
+            width: "500px",
+            data: {
+              title: "Mensaje",
+              message: "No existen secciones en el formato seleccionado!",
+            },
+          });
         }
       });
   }
